tracker age reverted to int, ageNano as uint64 in nanosecs

+++ libs/ofxCv/include/ofxCv/DNN.h
@@ -30,7 +30,7 @@ class Detection {
    float smoothHeight;
    float smoothCenterX;
    float smoothCenterY;
-   float age;
+   uint64_t age;
 
    bool operator<(Detection const& comp) const {
       return (age < comp.age);
